Updated menu and page profile to ignore irrelevant indicators
The menu ignores indicators if the child geographies don't have that indicator. This is to avoid a situation where a choropleth is not possible.
The page profile ignore an indicator if it has a dummy value i.e. -1. This is because dummy values must be added to the data to ensure that indicators that belong to child geographies but not the current geography are available in the menu