feat: adds first unit tests for FE
adding unit tests via jest to allow easy unit testing going forward
this is an example of a very simple unit test. it does not invole the UI
so it's just testing the business logic. in isolation.
I will add more tests involving the UI for unit tests as well as
integration tests, testing multiple parts.
needed to change the controller to use a "normal" function.
otherwise we would have needed to add more babel plugins

needed to change the .babelrc, not sure what the side effects of the env
deletion is just yet. need to investigate going forward.