Feature/add configurable graph domain (#136)
* Added additional configuration to the horizontal barchart
Can now specify the minX and maxX values

* Adding missing keys to chartConfiguration
previously the expectation was that all keys were present if a custom config was provided via the api

* Refactored chart code and allowed additional graph configuration
Some cleanup of chart.js. Removed some if statement that checked for percentage or value - this is now driven from the config. Also added the ability to set the domain of the axis. This allows users to for instance for the percentage graphs to range between 0 and 100

* Fixed test and replaced custom object merge function in utils with lodash
Lodash has a better deep merge function which deals with cases that the existing code didn't.

Co-authored-by: Adi Eyal <adi@openup.org.za>
Co-authored-by: Mila Frerichs <mila.frerichs@gmail.com>