Fixed default coordinates in config
Default coordinates can be set in the profile config. These will set the initial lat/lon/zoom when no geography is set in the hash fragment, i.e. /#geo:<geography> - previously this didn't not work as the default coordinates were set and then the map immediately panned to the default geography. This code keeps the shouldZoom state in the controller. shouldZoom is disabled when there is no geography in the fragment.