feat: use old chart-menu for interaction
use old chart menu for downloading a chart and changing the chart value
still to be done: width needs to be calculated properly, we now use a
magic number from the main panel. need to adjust that.
the values of the bars should always be visible, need to adjust the text
mark
still an improvement and gets us closer to releasing this feature