feat: add new features to dev tools
add the ability to change the default panel
add the ability to turn on a local server which mocks the API.
the API server uses miragejs for easier creation of data and responsing
the way our API works.
still a WIP but very nice for local dev when the BE is not ready yet.
will be expanded in the future with all API responses and the ability to
turn them on/off individually