chore: adjust indicator and chart
only send those items from indicator to chart that are actually needed.
adjust the update of the chart when the value type changes from
percentage to absolute
add pading to the right side for the data text (it overflows a lot, so
add a bit more padding to it)
switching between percentage and value works via the menu
filtering is commented out for now, needs to be adjusted.
check subindicators now uses data and creation of subindicators is not
needed for now