feat: use the actual values from the server
use either the local server or the actual one (once the changes are in
place) to test this out. removed the hard-coded values.
also removed and changed code that will need to be updated as well, as
we're now using the new data format for the population of the data
mapper as well
there is some work left for making sure when a geography changes that
the data is properly used
removed the on-the-fly getting of the indicator data for now until the
BE is ready. the data is available in the all_details request anyway for
now but this will change in the future