feat: add filtering back
filtering charts now works as well, most of the work was done in the
choropleth branch here:
https://github.com/OpenUpSA/wazimap-ng-ui/pull/290
I changed the callback structure a little bit and added the option to
not change the data but just send the filter.
maybe the changing needs to happen in a different place. the filter is
not a good place for it anyway